Tested types library
