+++ core-web/libs/types/jest.config.ts
@@ -1,0 +1,11 @@
+/* eslint-disable */
+export default {
+    displayName: 'types',
+    preset: '../../jest.preset.js',
+    testEnvironment: 'node',
+    transform: {
+        '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }]
+    },
+    moduleFileExtensions: ['ts', 'js', 'html'],
+    coverageDirectory: '../../coverage/libs/types'
+};
