+++ core-web/package.json
@@ -238,6 +238,7 @@
         "jest": "29.7.0",
         "jest-cli": "29.7.0",
         "jest-environment-jsdom": "29.7.0",
+        "jest-environment-node": "^29.7.0",
         "jest-fetch-mock": "^3.0.3",
         "jest-html-reporters": "^3.1.5",
         "jest-junit": "^16.0.0",
@@ -276,6 +277,7 @@
         "tslint-angular": "^3.0.3",
         "typedoc": "^0.25.4",
         "typescript": "5.4.5",
+        "verdaccio": "^5.0.4",
         "vite": "~5.0.0",
         "vite-plugin-dts": "~3.8.1",
         "vitest": "^1.3.1",
