+++ core-web/project.json
@@ -1,4 +1,14 @@
 {
     "name": "core-web",
-    "$schema": "node_modules/nx/schemas/project-schema.json"
+    "$schema": "node_modules/nx/schemas/project-schema.json",
+    "targets": {
+        "local-registry": {
+            "executor": "@nx/js:verdaccio",
+            "options": {
+                "port": 4873,
+                "config": ".verdaccio/config.yml",
+                "storage": "tmp/local-registry/storage"
+            }
+        }
+    }
 }
