+++ examples/astro/package.json
@@ -15,7 +15,6 @@
     "@astrojs/tailwind": "^5.1.0",
     "@astrojs/vercel": "^7.8.0",
     "@dotcms/client": "latest",
-    "@dotcms/react": "latest",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "astro": "^4.14.3",
